+++ include/zfile.h
@@ -50,7 +50,7 @@ extern struct zfile *zfile_fopen_load_zfile(struct zfile *f);
 extern uae_u8 *zfile_load_data(const TCHAR *name, const uae_u8 *data,int datalen, int *outlen);
 extern uae_u8 *zfile_load_file(const TCHAR *name, int *outlen);
 extern struct zfile *zfile_fopen_parent(struct zfile*, const TCHAR*, uae_u64 offset, uae_u64 size);
-extern uae_u8 *zfile_get_data_pointer(struct zfile *z, int *len);
+extern uae_u8 *zfile_get_data_pointer(struct zfile *z, size_t *len);
 
 extern int zfile_exists (const TCHAR *name);
 extern void zfile_fclose (struct zfile *z);
@@ -71,7 +71,7 @@ extern uae_u8 *zfile_getdata(struct zfile *z, uae_s64 offset, int len, int *outl
 extern void zfile_exit(void);
 extern int execute_command(TCHAR *);
 extern int zfile_iscompressed(struct zfile *z);
-extern int zfile_zcompress(struct zfile *dst, void *src, int size);
+extern int zfile_zcompress(struct zfile *dst, void *src, size_t size);
 extern int zfile_zuncompress(void *dst, int dstsize, struct zfile *src, int srcsize);
 extern int zfile_gettype(struct zfile *z);
 extern int zfile_zopen(const TCHAR *name, zfile_callback zc, void *user);
